refactor: create token panel component
