+++ packages/plugins/GoPlusSecurity/src/SNSAdaptor/components/SecurityPanel.tsx
@@ -1,53 +1,27 @@
-import { Link, Stack, Tooltip, Typography } from '@mui/material'
+import { Link, Stack, Typography } from '@mui/material'
 import { useI18N } from '../../locales'
 import { ExternalLink } from 'react-feather'
-import { makeStyles, usePortalShadowRoot } from '@masknet/theme'
+import { makeStyles } from '@masknet/theme'
 import { memo, useMemo, useState } from 'react'
 import { DefineMapping, SecurityMessageLevel, TokenSecurity } from './Common'
 import { SecurityMessages } from '../rules'
 import { RiskCard, RiskCardUI } from './RiskCard'
-import {
-    ERC20Token,
-    formatEthereumAddress,
-    resolveAddressLinkOnExplorer,
-    resolveTokenLinkOnExplorer,
-} from '@masknet/web3-shared-evm'
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown'
-import BigNumber from 'bignumber.js'
 import { useTheme } from '@mui/system'
 import { resolveGoLabLink } from '../../utils/helper'
-import parseInt from 'lodash-es/parseInt'
+import { TokenPanel } from './TokenPanel'
 
 interface TokenCardProps {
     tokenSecurity: TokenSecurity
 }
 
 const useStyles = makeStyles()((theme) => ({
-    root: {
-        width: '600px',
-    },
-    link: {},
-    card: {
-        borderColor: theme.palette.divider,
-        borderStyle: 'solid',
-        padding: theme.spacing(1.5),
-        borderRadius: 9,
-    },
-    subtitle: {
-        color: theme.palette.text.secondary,
-        fontWeight: 400,
-        fontSize: 14,
-    },
-    cardValue: {
-        color: theme.palette.text.primary,
-    },
-    tooltip: {
-        color: theme.palette.text.buttonText,
-        fontSize: 12,
-    },
     header: {
         fontWeight: 500,
         fontSize: 18,
+    },
+    root: {
+        width: '600px',
     },
     detectionCard: {
         backgroundColor: theme.palette.background.default,
@@ -60,17 +34,11 @@ const useStyles = makeStyles()((theme) => ({
     },
 }))
 
-const DEFAULT_PLACEHOLDER = '--'
-
-function formatTotalSupply(total?: number) {
-    if (!total) return DEFAULT_PLACEHOLDER
-    return new BigNumber(total).toFormat(3)
-}
-
 export const SecurityPanel = memo<TokenCardProps>(({ tokenSecurity }) => {
     const { classes } = useStyles()
     const t = useI18N()
     const theme = useTheme()
+
     const [isFold, setFold] = useState(false)
 
     const makeMessageList =
@@ -85,21 +53,6 @@ export const SecurityPanel = memo<TokenCardProps>(({ tokenSecurity }) => {
 
     const riskyFactors = makeMessageList.filter((x) => x.level === SecurityMessageLevel.High).length
     const attentionFactors = makeMessageList.filter((x) => x.level === SecurityMessageLevel.Medium).length
-
-    const totalSupply = usePortalShadowRoot((container) => {
-        return (
-            <Tooltip
-                PopperProps={{ container }}
-                arrow
-                title={
-                    <Typography color={(theme) => theme.palette.text.buttonText} className={classes.tooltip}>
-                        {tokenSecurity.total_supply}
-                    </Typography>
-                }>
-                <Typography className={classes.cardValue}>{formatTotalSupply(tokenSecurity.total_supply)}</Typography>
-            </Tooltip>
-        )
-    })
 
     const securityMessageLevel = useMemo(() => {
         if (riskyFactors) return SecurityMessageLevel.High
@@ -125,7 +78,6 @@ export const SecurityPanel = memo<TokenCardProps>(({ tokenSecurity }) => {
                             {t.more_details()}
                         </Typography>
                         <Link
-                            className={classes.link}
                             lineHeight="14px"
                             href={resolveGoLabLink(tokenSecurity.chainId, tokenSecurity.contract)}
                             target="_blank"
@@ -135,113 +87,8 @@ export const SecurityPanel = memo<TokenCardProps>(({ tokenSecurity }) => {
                         </Link>
                     </Stack>
                 </Stack>
-                {!isFold && (
-                    <Stack>
-                        <Stack direction="row" className={classes.card} spacing={2}>
-                            <Stack
-                                spacing={1.5}
-                                borderRadius={0.5}
-                                sx={{ background: DefineMapping[securityMessageLevel].bgColor }}
-                                width={128}
-                                height={128}
-                                justifyContent="center"
-                                alignItems="center">
-                                {DefineMapping[securityMessageLevel].icon(33)}
-                                <Typography color={DefineMapping[securityMessageLevel].titleColor} fontSize={14}>
-                                    {t[DefineMapping[securityMessageLevel].i18nKey]({ quantity: '' })}
-                                </Typography>
-                            </Stack>
-                            <Stack height={128} justifyContent="space-between" flex={1}>
-                                <Stack direction="row" justifyContent="space-between">
-                                    <Typography className={classes.subtitle}>{t.token_info_token_name()}</Typography>
-                                    <Typography className={classes.cardValue}>
-                                        {tokenSecurity.token_symbol}({tokenSecurity.token_name})
-                                    </Typography>
-                                </Stack>
-                                <Stack direction="row" justifyContent="space-between">
-                                    <Typography className={classes.subtitle}>
-                                        {t.token_info_token_contract_address()}
-                                    </Typography>
-                                    <Stack display="inline-flex" direction="row" alignItems="center" spacing={0.625}>
-                                        <Typography className={classes.cardValue}>
-                                            {tokenSecurity.contract
-                                                ? formatEthereumAddress(tokenSecurity.contract, 4)
-                                                : DEFAULT_PLACEHOLDER}
-                                        </Typography>
-                                        <Link
-                                            className={classes.link}
-                                            lineHeight="14px"
-                                            href={resolveTokenLinkOnExplorer({
-                                                chainId: parseInt(tokenSecurity.chainId),
-                                                address: tokenSecurity.contract,
-                                            } as ERC20Token)}
-                                            target="_blank"
-                                            title={t.token_info_token_contract_address()}
-                                            rel="noopener noreferrer">
-                                            <ExternalLink color={theme.palette.text.strong} size={14} />
-                                        </Link>
-                                    </Stack>
-                                </Stack>
-                                <Stack direction="row" justifyContent="space-between">
-                                    <Typography className={classes.subtitle}>
-                                        {t.token_info_contract_creator()}
-                                    </Typography>
-                                    <Stack display="inline-flex" direction="row" alignItems="center" spacing={0.625}>
-                                        <Typography className={classes.cardValue}>
-                                            {tokenSecurity.creator_address
-                                                ? formatEthereumAddress(tokenSecurity.creator_address ?? '', 4)
-                                                : DEFAULT_PLACEHOLDER}
-                                        </Typography>
-                                        {tokenSecurity.creator_address && (
-                                            <Link
-                                                className={classes.link}
-                                                lineHeight="14px"
-                                                href={resolveAddressLinkOnExplorer(
-                                                    parseInt(tokenSecurity.chainId),
-                                                    tokenSecurity.creator_address,
-                                                )}
-                                                target="_blank"
-                                                title={t.token_info_contract_creator()}
-                                                rel="noopener noreferrer">
-                                                <ExternalLink color={theme.palette.text.strong} size={14} />
-                                            </Link>
-                                        )}
-                                    </Stack>
-                                </Stack>
-                                <Stack direction="row" justifyContent="space-between">
-                                    <Typography className={classes.subtitle}>
-                                        {t.token_info_contract_owner()}
-                                    </Typography>
-                                    <Stack display="inline-flex" direction="row" alignItems="center" spacing={0.625}>
-                                        <Typography className={classes.cardValue}>
-                                            {tokenSecurity.owner_address
-                                                ? formatEthereumAddress(tokenSecurity.owner_address ?? '', 4)
-                                                : DEFAULT_PLACEHOLDER}
-                                        </Typography>
-                                        {tokenSecurity.owner_address && (
-                                            <Link
-                                                className={classes.link}
-                                                lineHeight="14px"
-                                                href={resolveAddressLinkOnExplorer(
-                                                    parseInt(tokenSecurity.chainId),
-                                                    tokenSecurity.owner_address,
-                                                )}
-                                                target="_blank"
-                                                title={t.token_info_contract_owner()}
-                                                rel="noopener noreferrer">
-                                                <ExternalLink color={theme.palette.text.strong} size={14} />
-                                            </Link>
-                                        )}
-                                    </Stack>
-                                </Stack>
-                                <Stack direction="row" justifyContent="space-between">
-                                    <Typography className={classes.subtitle}>{t.token_info_total_supply()}</Typography>
-                                    {totalSupply}
-                                </Stack>
-                            </Stack>
-                        </Stack>
-                    </Stack>
-                )}
+
+                {!isFold && <TokenPanel tokenSecurity={tokenSecurity} securityMessageLevel={securityMessageLevel} />}
             </Stack>
             <Stack spacing={2}>
                 <Stack direction="row" alignItems="center" spacing={3.5}>
